Refactor:공간 기록 관련 컴포넌트 아토믹 디자인으로 정리
