+++ app/feature/home/components/HomeSlider/molecules/SliderHashTag.tsx
