+++ app/feature/home/components/HomeSlider/slick.css
