+++ app/feature/record/components/RecordComplete/organisms/RecordComplete.tsx
@@ -10,6 +10,7 @@ interface RecordCompleteProps {
   hasSomeData: boolean;
 }
 
+//Organism
 export default function RecordComplete({
   spaceId,
   handleIndicatorIndex,
