+++ app/feature/record/components/UserRecord/skeleton/RecordMyPageSkeleton.tsx
