+++ app/feature/search/hooks/SearchText/useTextSearchBar.tsx
